Update errors structure in toolkit
