+++ toolkit/src/errors/account.rs
@@ -67,51 +67,3 @@ impl From<std::io::Error> for AddressError {
         AddressError::Crate("std::io", format!("{:?}", error))
     }
 }
-
-#[derive(Debug, Error)]
-pub enum SignatureError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-}
-
-impl From<hex::FromHexError> for SignatureError {
-    fn from(error: hex::FromHexError) -> Self {
-        SignatureError::Crate("hex", format!("{:?}", error))
-    }
-}
-
-impl From<snarkos_errors::objects::account::AccountError> for SignatureError {
-    fn from(error: snarkos_errors::objects::account::AccountError) -> Self {
-        SignatureError::Crate("snarkos_errors::objects::account", format!("{:?}", error))
-    }
-}
-
-impl From<snarkos_errors::algorithms::signature::SignatureError> for SignatureError {
-    fn from(error: snarkos_errors::algorithms::signature::SignatureError) -> Self {
-        SignatureError::Crate("snarkos_errors::algorithms::signature", format!("{:?}", error))
-    }
-}
-
-impl From<std::io::Error> for SignatureError {
-    fn from(error: std::io::Error) -> Self {
-        SignatureError::Crate("std::io", format!("{:?}", error))
-    }
-}
-
-#[derive(Debug, Error)]
-pub enum RecordError {
-    #[error("{}: {}", _0, _1)]
-    Crate(&'static str, String),
-}
-
-impl From<snarkos_errors::dpc::DPCError> for RecordError {
-    fn from(error: snarkos_errors::dpc::DPCError) -> Self {
-        RecordError::Crate("snarkos_errors::dpc", format!("{:?}", error))
-    }
-}
-
-impl From<std::io::Error> for RecordError {
-    fn from(error: std::io::Error) -> Self {
-        RecordError::Crate("std::io", format!("{:?}", error))
-    }
-}
